config: Add Accounts button to main menu
Loading this pack can take a while. It would suck to have to relaunch
the game due to expired session token. Add Accounts button which was
previously hidden by the custom title screen. While we're here, also add
more screens to the customizable menu list so that they all have the
same background.