tweaks: Rearrange Accessory slots
tweaks: Move Elytra to Cape slot

Aside for cosmetic purposes, the Cape slot is lacking in functionality.
Also, Elytra is thematically more of a cape than a back equipment. Move
it.

tweaks: Add Bundle to Belt Slot

Coordinates (and weather and time) are only shown as HUD components when
the appropriate items (plural) are equipped. One way to minimize the
amount of inventory slots used is to add the items to a Bundle. Allow
equipping Bundle to the Belt Slot so the consumed slot is an Accessory
slot rather than the main inventory slot.